Add Refiner option to ImageData component
The change introduces a new 'refiner' option to the ImageData component with its associated dispatch action. The update allows users to select the 'refiner' from selectable properties and dispatches the 'setRefiner' action to set the refiner state. This also includes providing an interface for 'checkpoint' and 'switchAt' properties of the refiner.